# go-order-lite/docs/swagger.yaml
basePath: /api
definitions:
  dto.CreateOrderRequest:
    properties:
      amount:
        type: integer
    required:
    - amount
    type: object
  dto.LoginRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.OrderResponse:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      user_id:
        type: integer
    type: object
  handler.RegisterRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: phantomor@163.com
    name: Pvr1sC
  description: 一个用于学习的订单系统后端（Gin + JWT + Redis + MySQL）
  termsOfService: http://example.com
  title: Go Order Lite API
  version: "1.0"
paths:
  /../login:
    post:
      consumes:
      - application/json
      description: 用户名密码登录，返回 JWT Token
      parameters:
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - auth
  /../register:
    post:
      consumes:
      - application/json
      description: 注册新用户，需提供用户名和密码
      parameters:
      - description: 注册参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - user
  /order:
    get:
      consumes:
      - application/json
      description: 查询当前登录用户的订单（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询我的订单列表
      tags:
      - order
    post:
      consumes:
      - application/json
      description: 用户创建订单（JWT 鉴权 + RequestID 幂等）
      parameters:
      - description: 请求唯一ID（用于防重复下单）
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: 创建订单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建订单
      tags:
      - order
  /order/{id}/cancel:
    get:
      consumes:
      - application/json
      description: 用户主动取消订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 取消订单
      tags:
      - order
  /order/{id}/pay:
    get:
      consumes:
      - application/json
      description: 模拟用户支付订单
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 支付订单
      tags:
      - order
  /user/info:
    get:
      consumes:
      - application/json
      description: 通过 JWT Token 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"